--- internal/api/attestationconfigapi/cli/delete_test.go
@@ -1,6 +1,0 @@
-/*
-Copyright (c) Edgeless Systems GmbH
-
-SPDX-License-Identifier: AGPL-3.0-only
-*/
-package main
